Split off the parts we don't want to share.
